ingestion-service/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
numpy
python-multipart